changed Double to FLoat in LongLatDTO

+++ src/main/java/be/rdhaese/packetdelivery/dto/LongLatDTO.java
@@ -6,14 +6,14 @@ package be.rdhaese.packetdelivery.dto;
  * @author Robin D'Haese
  */
 public class LongLatDTO {
-    private Float longitude;
-    private Float latitude;
+    private Double longitude;
+    private Double latitude;
 
     public LongLatDTO() {
 
     }
 
-    public LongLatDTO(Float latitude, Float longitude) {
+    public LongLatDTO(Double latitude, Double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
@@ -38,19 +38,19 @@ public class LongLatDTO {
         return result;
     }
 
-    public Float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
-    public Float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 }
